Improvements to example 6
git-svn-id: https://svn.mcs.anl.gov/repos/exm/sfw/stc/trunk@8036 dc4e9af1-7f46-4ead-bba6-71afc04862de

+++ docs/examples/6/f.h
@@ -1,2 +1,2 @@
 
-int f(MPI_Comm comm, int k);
+double f(MPI_Comm comm, int k);
